LayoutTests/imported/w3c/web-platform-tests/xhr/idlharness.any-expected.txt

CONSOLE MESSAGE: line 5: ReferenceError: Can't find variable: idl_test

FAIL Untitled ReferenceError: Can't find variable: idl_test
